Set fullpage as directive, not component
