+++ ui/src/app/cohort-review/cohort-review.module.ts
@@ -5,7 +5,7 @@ import {RouterModule} from '@angular/router';
 import {ClarityModule} from 'clarity-angular';
 
 import {CohortReviewComponent} from './cohort-review/cohort-review.component';
-import {FullpageAppComponent} from './fullpage-app/fullpage-app.component';
+import {FullpageDirective} from './directives';
 import {CohortResolverGuard, ReviewResolverGuard} from './guards';
 import {OverviewComponent} from './overview/overview.component';
 import {SubjectDetailComponent} from './subject-detail/subject-detail.component';
@@ -39,7 +39,7 @@ const routes = [{
     CohortReviewComponent,
     OverviewComponent,
     SubjectDetailComponent,
-    FullpageAppComponent,
+    FullpageDirective,
   ],
   providers: [
     CohortResolverGuard,
