+++ ui/src/app/cohort-review/cohort-review/cohort-review.component.css
@@ -1,11 +1,3 @@
-/* Structural Styles */
-.cohort-review-wrapper {
-  background-color: rgb(250, 250, 250);
-  padding: 1rem;
-  margin-bottom: 1rem;
-  position: relative;
-}
-
 .sidebar {
   height: 100%;
   width: 0;
