+++ ui/src/app/cohort-review/cohort-review/cohort-review.component.html
@@ -1,4 +1,4 @@
-<app-fullpage-app>
+<div appFullpage>
   <div *ngIf="loading" class="spinner">Loading...</div>
 
   <!-- Main subject navigation sidebar -->
@@ -133,4 +133,4 @@
       </button>
     </div>
   </clr-modal>
-</app-fullpage-app>
+</div>
