+++ ui/src/app/cohort-review/directives/index.ts
@@ -1,0 +1,1 @@
+export * from './fullpage';
