--- ui/src/app/cohort-review/fullpage-app/fullpage-app.component.css
@@ -1,5 +1,0 @@
-.fullpage-app-wrapper {
-  background-color: rgb(250, 250, 250);
-  padding: 1rem;
-  margin-bottom: 1rem;
-}
