--- ui/src/app/cohort-review/fullpage-app/fullpage-app.component.html
@@ -1,3 +1,0 @@
-<div #wrapper class="fullpage-app-wrapper">
-  <ng-content></ng-content>
-</div>
